refactor(game): use implicit index imports
