+++ src/game/reducers/shuffle-deck/index.ts
@@ -1,7 +1,7 @@
 import shuffle from 'lodash.shuffle'
 
 import { IGame, IPlayer } from '../../types'
-import { updateDeck } from '../update-deck/index'
+import { updateDeck } from '../update-deck'
 
 export const shuffleDeck = (game: IGame, playerId: IPlayer['id']): IGame => {
   const { deck } = game.table.players[playerId]
